code/Bigquery_BDCC-main/requirements.txt
Flask
gunicorn
google-cloud-bigquery
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2